requirements.txt
Flask
langchain
python-dotenv
langchain-google-genai 
google-generativeai


